Refactor test program.
