+++ docs/ERD.md
@@ -1,9 +1,11 @@
 # PROJECT
+
 > Generated by [`prisma-markdown`](https://github.com/samchon/prisma-markdown)
 
 - [Articles](#articles)
 
 ## Articles
+
 ```mermaid
 erDiagram
 "attachment_files" {
@@ -63,6 +65,7 @@ erDiagram
 ```
 
 ### `attachment_files`
+
 Attachment File.
 
 Every attachment files that are managed in current system.
@@ -70,20 +73,22 @@ Every attachment files that are managed in current system.
 For reference, it is possible to omit one of file name or extension like 
 `.gitignore` or `README` case, but not possible to omit both of them.
 
-**Properties**
-  - `id`: 
-  - `name`
-    > File name, except extension.
-    > 
-    > If there's file `.gitignore`, then its name is an empty string.
-  - `extension`
-    > Extension.
-    > 
-    > Possible to omit like `README` case.
-  - `url`: URL path of the real file.
-  - `created_at`: Creation time of file.
+Properties as follows:
+
+- `id`:
+- `name`
+  > File name, except extension.
+  >
+  > If there's file `.gitignore`, then its name is an empty string.
+- `extension`
+  > Extension.
+  >
+  > Possible to omit like `README` case.
+- `url`: URL path of the real file.
+- `created_at`: Creation time of file.
 
 ### `bbs_articles`
+
 Article entity.
 
 `bbs_articles` is a super-type entity of all kinds of articles in the 
@@ -105,16 +110,18 @@ the article is designed in this structure.
 
 In other words, to keep evidence, and prevent fraud.
 
-**Properties**
-  - `id`: 
-  - `created_at`: Creation time of article.
-  - `deleted_at`
-    > Deletion time of article.
-    > 
-    > To keep evidence, do not delete the article, but just mark it as 
-    > deleted.
+Properties as follows:
+
+- `id`:
+- `created_at`: Creation time of article.
+- `deleted_at`
+  > Deletion time of article.
+  >
+  > To keep evidence, do not delete the article, but just mark it as
+  > deleted.
 
 ### `bbs_article_snapshots`
+
 Snapshot of article.
 
 `bbs_article_snapshots` is a snapshot entity that contains the contents of
@@ -122,21 +129,23 @@ the article, as mentioned in [bbs_articles](#bbs_articles), the contents of the
 article are separated from the article record to keep evidence and prevent 
 fraud.
 
-**Properties**
-  - `id`: 
-  - `bbs_article_id`: Belong article's [bbs_articles.id](#bbs_articles)
-  - `format`
-    > Format of body.
-    > 
-    > Same meaning with extension like `html`, `md`, `txt`.
-  - `title`: Title of article.
-  - `body`: Content body of article.
-  - `created_at`
-    > Creation time of record.
-    > 
-    > It means creation time or update time or article.
+Properties as follows:
+
+- `id`:
+- `bbs_article_id`: Belong article's [bbs_articles.id](#bbs_articles)
+- `format`
+  > Format of body.
+  >
+  > Same meaning with extension like `html`, `md`, `txt`.
+- `title`: Title of article.
+- `body`: Content body of article.
+- `created_at`
+  > Creation time of record.
+  >
+  > It means creation time or update time or article.
 
 ### `bbs_article_snapshot_files`
+
 Attachment file of article snapshot.
 
 `bbs_article_snapshot_files` is an entity that shapes the attached files of
@@ -148,13 +157,15 @@ resolve the M: N relationship between [bbs_article_snapshots](#bbs_article_snaps
 files, it has an additional `sequence` attribute, which we will continue to
 see in this documents.
 
-**Properties**
-  - `id`: 
-  - `bbs_article_snapshot_id`: Belonged snapshot's [bbs_article_snapshots.id](#bbs_article_snapshots)
-  - `attachment_file_id`: Belonged file's [attachment_files.id](#attachment_files)
-  - `sequence`: Sequence of attachment file in the snapshot.
+Properties as follows:
+
+- `id`:
+- `bbs_article_snapshot_id`: Belonged snapshot's [bbs_article_snapshots.id](#bbs_article_snapshots)
+- `attachment_file_id`: Belonged file's [attachment_files.id](#attachment_files)
+- `sequence`: Sequence of attachment file in the snapshot.
 
 ### `bbs_article_comments`
+
 Comment written on an article.
 
 `bbs_article_comments` is an entity that shapes the comments written on an
@@ -169,21 +180,23 @@ the comment is modified.
 Also, `bbs_article_comments` is expressing the relationship of the 
 hierarchical reply structure through the `parent_id` attribute.
 
-**Properties**
-  - `id`: 
-  - `bbs_article_id`: Belonged article's [bbs_articles.id](#bbs_articles)
-  - `parent_id`
-    > Parent comment's [bbs_article_comments.id](#bbs_article_comments)
-    > 
-    > Used to express the hierarchical reply structure.
-  - `created_at`: Creation time of comment.
-  - `deleted_at`
-    > Deletion time of comment.
-    > 
-    > Do not allow to delete the comment, but just mark it as deleted, 
-    > to keep evidence.
+Properties as follows:
+
+- `id`:
+- `bbs_article_id`: Belonged article's [bbs_articles.id](#bbs_articles)
+- `parent_id`
+  > Parent comment's [bbs_article_comments.id](#bbs_article_comments)
+  >
+  > Used to express the hierarchical reply structure.
+- `created_at`: Creation time of comment.
+- `deleted_at`
+  > Deletion time of comment.
+  >
+  > Do not allow to delete the comment, but just mark it as deleted,
+  > to keep evidence.
 
 ### `bbs_article_comment_snapshots`
+
 Snapshot of comment.
 
 `bbs_article_comment_snapshots` is a snapshot entity that contains the 
@@ -192,31 +205,34 @@ contents of the comment.
 As mentioned in [bbs_article_comments](#bbs_article_comments), designed to keep evidence 
 and prevent fraud.
 
-**Properties**
-  - `id`: 
-  - `bbs_article_comment_id`: Belonged article's [bbs_article_comments.id](#bbs_article_comments)
-  - `format`
-    > Format of content body.
-    > 
-    > Same meaning with extension like `html`, `md`, `txt`.
-  - `body`: Content body of comment.
-  - `created_at`
-    > Creation time of record.
-    > 
-    > It means creation time or update time or comment.
+Properties as follows:
+
+- `id`:
+- `bbs_article_comment_id`: Belonged article's [bbs_article_comments.id](#bbs_article_comments)
+- `format`
+  > Format of content body.
+  >
+  > Same meaning with extension like `html`, `md`, `txt`.
+- `body`: Content body of comment.
+- `created_at`
+  > Creation time of record.
+  >
+  > It means creation time or update time or comment.
 
 ### `bbs_article_comment_snapshot_files`
+
 Attachment file of comment snapshot.
 
 `bbs_article_comment_snapshot_files` is an entity resolving the M:N 
 relationship between [bbs_article_comment_snapshots](#bbs_article_comment_snapshots) and 
 [attachment_files](#attachment_files) tables.
 
-**Properties**
-  - `id`: 
-  - `bbs_article_comment_snapshot_id`: Belonged snapshot's [bbs_article_comment_snapshots.id](#bbs_article_comment_snapshots)
-  - `attachment_file_id`: Belonged file's [attachment_files.id](#attachment_files)
-  - `sequence`
-    > Sequence order.
-    > 
-    > Sequence order of the attached file in the belonged snapshot.
+Properties as follows:
+
+- `id`:
+- `bbs_article_comment_snapshot_id`: Belonged snapshot's [bbs_article_comment_snapshots.id](#bbs_article_comment_snapshots)
+- `attachment_file_id`: Belonged file's [attachment_files.id](#attachment_files)
+- `sequence`
+  > Sequence order.
+  >
+  > Sequence order of the attached file in the belonged snapshot.
