+++ src/api/functional/monitors/health/index.ts
@@ -19,7 +19,7 @@ import typia from "typia";
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 export async function get(connection: IConnection): Promise<void> {
-  return !!connection.simulate
+  return true === connection.simulate
     ? get.simulate(connection)
     : PlainFetcher.fetch(connection, {
         ...get.METADATA,
@@ -40,13 +40,8 @@ export namespace get {
   } as const;
 
   export const path = () => "/monitors/health";
-  export const random = (g?: Partial<typia.IRandomGenerator>): void =>
-    typia.random<void>(g);
-  export const simulate = (connection: IConnection): void => {
-    return random(
-      "object" === typeof connection.simulate && null !== connection.simulate
-        ? connection.simulate
-        : undefined,
-    );
+  export const random = (): void => typia.random<void>();
+  export const simulate = (_connection: IConnection): void => {
+    return random();
   };
 }
