+++ src/api/functional/monitors/performance/index.ts
@@ -24,7 +24,7 @@ import type { IPerformance } from "../../../structures/monitors/IPerformance";
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 export async function get(connection: IConnection): Promise<get.Output> {
-  return !!connection.simulate
+  return true === connection.simulate
     ? get.simulate(connection)
     : PlainFetcher.fetch(connection, {
         ...get.METADATA,
@@ -47,13 +47,8 @@ export namespace get {
   } as const;
 
   export const path = () => "/monitors/performance";
-  export const random = (g?: Partial<typia.IRandomGenerator>): IPerformance =>
-    typia.random<IPerformance>(g);
-  export const simulate = (connection: IConnection): Output => {
-    return random(
-      "object" === typeof connection.simulate && null !== connection.simulate
-        ? connection.simulate
-        : undefined,
-    );
+  export const random = (): IPerformance => typia.random<IPerformance>();
+  export const simulate = (_connection: IConnection): Output => {
+    return random();
   };
 }
