+++ src/api/functional/monitors/system/index.ts
@@ -24,7 +24,7 @@ import type { ISystem } from "../../../structures/monitors/ISystem";
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 export async function get(connection: IConnection): Promise<get.Output> {
-  return !!connection.simulate
+  return true === connection.simulate
     ? get.simulate(connection)
     : PlainFetcher.fetch(connection, {
         ...get.METADATA,
@@ -47,14 +47,9 @@ export namespace get {
   } as const;
 
   export const path = () => "/monitors/system";
-  export const random = (g?: Partial<typia.IRandomGenerator>): ISystem =>
-    typia.random<ISystem>(g);
-  export const simulate = (connection: IConnection): Output => {
-    return random(
-      "object" === typeof connection.simulate && null !== connection.simulate
-        ? connection.simulate
-        : undefined,
-    );
+  export const random = (): ISystem => typia.random<ISystem>();
+  export const simulate = (_connection: IConnection): Output => {
+    return random();
   };
 }
 
@@ -67,7 +62,7 @@ export namespace get {
 export async function internal_server_error(
   connection: IConnection,
 ): Promise<void> {
-  return !!connection.simulate
+  return true === connection.simulate
     ? internal_server_error.simulate(connection)
     : PlainFetcher.fetch(connection, {
         ...internal_server_error.METADATA,
@@ -88,14 +83,9 @@ export namespace internal_server_error {
   } as const;
 
   export const path = () => "/monitors/system/internal_server_error";
-  export const random = (g?: Partial<typia.IRandomGenerator>): void =>
-    typia.random<void>(g);
-  export const simulate = (connection: IConnection): void => {
-    return random(
-      "object" === typeof connection.simulate && null !== connection.simulate
-        ? connection.simulate
-        : undefined,
-    );
+  export const random = (): void => typia.random<void>();
+  export const simulate = (_connection: IConnection): void => {
+    return random();
   };
 }
 
@@ -108,7 +98,7 @@ export namespace internal_server_error {
 export async function uncaught_exception(
   connection: IConnection,
 ): Promise<void> {
-  return !!connection.simulate
+  return true === connection.simulate
     ? uncaught_exception.simulate(connection)
     : PlainFetcher.fetch(connection, {
         ...uncaught_exception.METADATA,
@@ -129,13 +119,8 @@ export namespace uncaught_exception {
   } as const;
 
   export const path = () => "/monitors/system/uncaught_exception";
-  export const random = (g?: Partial<typia.IRandomGenerator>): void =>
-    typia.random<void>(g);
-  export const simulate = (connection: IConnection): void => {
-    return random(
-      "object" === typeof connection.simulate && null !== connection.simulate
-        ? connection.simulate
-        : undefined,
-    );
+  export const random = (): void => typia.random<void>();
+  export const simulate = (_connection: IConnection): void => {
+    return random();
   };
 }
